requirements.txt
streamlit
langchain>=0.1.13
langchain-openai>=0.1.8
langchain-chroma>=0.1.0
chromadb>=0.4.24
openai>=1.0.0
ragas>=0.2.15
pandas>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
beautifulsoup4>=4.12.0
requests>=2.31.0
ipython
jupyter 